Removed bookkeepers, height became uint64

+++ core/store/common/store.go
@@ -2,6 +2,7 @@ package common
 
 import (
 	"errors"
+
 	"github.com/eywa-protocol/chain/common"
 	"github.com/eywa-protocol/chain/core/states"
 	"github.com/eywa-protocol/chain/native/event"
@@ -69,7 +70,7 @@ type EventStore interface {
 	//SaveEventNotifyByTx save event notify gen by smart contract execution
 	SaveEventNotifyByTx(txHash common.Uint256, notify *event.ExecuteNotify) error
 	//Save transaction hashes which have event notify gen
-	SaveEventNotifyByBlock(height uint32, txHashs []common.Uint256) error
+	SaveEventNotifyByBlock(height uint64, txHashs []common.Uint256) error
 	//GetEventNotifyByTx return event notify by transaction hash
 	GetEventNotifyByTx(txHash common.Uint256) (*event.ExecuteNotify, error)
 	//Commit event notify to store
